Renamed pwndbg.file to pwndbg.gdblib.file
