+++ pwndbg/commands/mprotect.py
@@ -7,7 +7,7 @@ from pwnlib import asm
 import pwndbg.chain
 import pwndbg.commands
 import pwndbg.enhance
-import pwndbg.file
+import pwndbg.gdblib.file
 import pwndbg.lib.which
 import pwndbg.wrappers.checksec
 import pwndbg.wrappers.readelf
